Website Update
Completed steps 1 - 7 by creating the setupCart function that sets up the event handlers for the "Add to Order" buttons on the web page, and creating the addItem function that adds the food item associated with the "Add to Order" button to the shopping cart, keeping track of the number of items of each product ordered by the customer. Also added comments.